Oppdatert package names og fiks i controllers
