+++ apps/tp-service/src/main/java/no/nav/testnav/apps/tpservice/consumer/command/hodejegeren/GetAlleLevendeCommand.java
@@ -1,0 +1,29 @@
+package no.nav.testnav.apps.tpservice.consumer.command.hodejegeren;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.core.ParameterizedTypeReference;
+import org.springframework.web.reactive.function.client.WebClient;
+
+import java.util.List;
+import java.util.concurrent.Callable;
+
+@RequiredArgsConstructor
+public class GetAlleLevendeCommand implements Callable<List<String>> {
+
+    private static final ParameterizedTypeReference<List<String>> RESPONSE_TYPE = new ParameterizedTypeReference<>() {
+    };
+    private final Long avspillergruppeId;
+    private final WebClient webClient;
+
+    @Override
+    public List<String> call() {
+        return webClient.get()
+                .uri(builder ->
+                        builder.path("/api/v1/alle-levende-identer/{avspillergruppeId}")
+                                .build(avspillergruppeId)
+                )
+                .retrieve()
+                .bodyToMono(RESPONSE_TYPE)
+                .block();
+    }
+}
