--- apps/tp-service/src/main/java/no/nav/testnav/apps/tpservice/consumer/rs/credential/SyntTpProperties.java
@@ -1,11 +1,0 @@
-package no.nav.testnav.apps.tpservice.consumer.rs.credential;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-
-import no.nav.testnav.libs.securitycore.domain.ServerProperties;
-
-@Configuration
-@ConfigurationProperties(prefix = "consumers.synt-tp-gcp")
-public class SyntTpProperties extends ServerProperties {
-}
