--- apps/tp-service/src/main/java/no/nav/testnav/apps/tpservice/consumer/rs/request/LagreTpForholdRequest.java
@@ -1,20 +1,0 @@
-package no.nav.testnav.apps.tpservice.consumer.rs.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.util.List;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class LagreTpForholdRequest {
-
-    private List<String> miljoer;
-
-    private String fnr;
-    private String ordning;
-}
