--- apps/tp-service/src/main/java/no/nav/testnav/apps/tpservice/provider/rs/request/OrkestreringRequest.java
@@ -1,24 +1,0 @@
-package no.nav.testnav.apps.tpservice.provider.rs.request;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.validation.constraints.NotNull;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class OrkestreringRequest {
-
-    @NotNull
-    @JsonProperty("avspillergruppeId")
-    private Long avspillergruppeId;
-    @NotNull
-    @JsonProperty("miljoe")
-    private String miljoe;
-
-}
