+++ apps/tp-service/src/main/java/no/nav/testnav/apps/tpservice/service/SyntService.java
@@ -3,8 +3,8 @@ package no.nav.testnav.apps.tpservice.service;
 import io.micrometer.core.annotation.Timed;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import no.nav.testnav.apps.tpservice.consumer.rs.HodejegerenConsumer;
-import no.nav.testnav.apps.tpservice.consumer.rs.TpSyntConsumer;
+import no.nav.testnav.apps.tpservice.consumer.HodejegerenConsumer;
+import no.nav.testnav.apps.tpservice.consumer.TpSyntConsumer;
 import no.nav.testnav.apps.tpservice.database.models.HistorikkComposityKey;
 import no.nav.testnav.apps.tpservice.database.models.TForholdYtelseHistorikk;
 import no.nav.testnav.apps.tpservice.database.models.TPerson;
@@ -20,7 +20,7 @@ import no.nav.testnav.apps.tpservice.domain.Person;
 import no.nav.testnav.apps.tpservice.domain.TpSaveInHodejegerenRequest;
 import no.nav.testnav.apps.tpservice.domain.Ytelse;
 import no.nav.testnav.apps.tpservice.domain.FulltForhold;
-import no.nav.testnav.apps.tpservice.provider.rs.request.SyntetiseringsRequest;
+import no.nav.testnav.apps.tpservice.provider.request.SyntetiseringsRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.HttpServerErrorException;
